chore: 코드 분할
각 폴더마다 수행하는 기능이 다르게 나눠놨으므로, 이를 위해 코드들을 옮기고, 파일들을 생성함
