+++ src/shared/ui/login/authTermStytle.ts
@@ -1,0 +1,22 @@
+export const styles = {
+  container: {
+    mt: 5,
+  },
+  loadingBox: {
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    minHeight: '50vh',
+  },
+  card: {
+    mb: 2,
+    boxShadow: 2,
+  },
+  termText: {
+    mt: 1,
+  },
+  submitButton: {
+    mt: 2,
+    width: '100%',
+  },
+}
